refactor(domain): migliora modelli di dominio SRD
Aggiorna e migliora i modelli di dominio per tutte le entità SRD
includendo strutture dati più robuste e terminologia italiana.

- Migliora modello Animali con statistiche strutturate
- Aggiorna Armatura con campi costo/peso strutturati
- Ottimizza modelli Classe, Incantesimo, Mostro
- Consolida logica comune di dominio

+++ internal/domain/incantesimo.go
@@ -32,8 +32,8 @@ type TempoLancio struct {
 
 // Distanza
 type Distanza struct {
-	Valore int    `json:"valore" bson:"valore"`
-	Unita  string `json:"unita"  bson:"unita"` // "ft","m","km","mile"
+	Valore float64 `json:"valore" bson:"valore"`
+	Unita  string  `json:"unita"  bson:"unita"` // "ft","m","km","mile"
 }
 
 // Gittata
